requirements.txt
google-genai>=0.3.0
litellm>=1.0.0
faiss-cpu>=1.7.0
pypdf>=3.0.0
langchain-text-splitters>=0.2.0
python-dotenv>=1.0.0
numpy>=1.24.0
